# openAPI/openapi.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Shop
  description: Shortener сервис, позволяющий пользователю генерировать для произвольного URL’a его короткую версию.
  contact:
    name: send to email
    email: p.v.nedoshivin@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  /stat/{shortLink}:
    get:
      tags:
      - link
      summary: Shows statistics of clicks on a short link
      operationId: GetShortLink
      parameters:
      - name: shortLink
        in: path
        description: short link conversion statistics to return
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'
        404:
          description: Short link not found
          content: {}
      security:
        - api_key: []
components:
  schemas:
...
    Link:
      type: object
      properties:
        id:
          type: integer
          format: int64
        client_id:
          type: integer
          format: int64
        short_link:
          type: string
        full_link:
          type: string
        count:
          type: integer
          format: int64
        created_at:
          type: string
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header